+++ src/User/User.php
@@ -7,6 +7,7 @@ use Illuminate\Contracts\Auth\MustVerifyEmail as MustVerifyEmailContract;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Support\Str;
+use Illuminate\Validation\ValidationException;
 use Laravel\Passport\HasApiTokens;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 
@@ -63,5 +64,22 @@ class User extends Authenticatable implements MustVerifyEmailContract
     public function routeNotificationForMail()
     {
         return $this->controlUser()->data()->email();
+    }
+
+    /**
+     * Get the email address that should be used for verification.
+     *
+     * @return string
+     * @throws ValidationException
+     */
+    public function getEmailForVerification()
+    {
+        $email = $this->controlUser()->data()->email();
+        if(!$email) {
+            throw ValidationException::withMessages([
+                'identifier' => 'Your email has not been set.'
+            ]);
+        }
+        return $email;
     }
 }